Work so far.

+++ Tuna/SdFile.cpp
@@ -64,7 +64,7 @@ int16_t SdFile::write(const void* buf, uint16_t nbyte) {
  * Use writeError to check for errors.
  */
 #if ARDUINO >= 100
-  size_t SdFile::write(uint8_t b) {
+uint8_t SdFile::write(uint8_t b) {
     return SdBaseFile::write(&b, 1);
   }
 #else
